fix: rework environment mocks, organize fixtures and mocks
