+++ test/wtr/fixtures/sidekick-config.js
