+++ test/wtr/mocks/i18n.js
